Zap deposit test
