+++ src/config/zap/optimism.tsx
@@ -1,10 +1,11 @@
 export const zaps = [
-  // {
-  // velodrome
-  //zapAddress: '0x2C3d263b560b2700fcf19e78f2EBD4d59EB7b3c5',
-  //ammRouter: '0xa132DAB612dB5cB9fC9Ac426A0Cc215A3423F9c9',
-  //ammFactory: '0x25CbdDb98b35ab1FF77413456B31EC81A6B6B746',
-  //ammPairInitHash: '0xe31da4209ffcce713230a74b5287fa8ec84797c9e77e1f7cfeccea015cdc97ea',
-  //lpProviderFee: 0.0,
-  // },
+  {
+    // velodrome
+    type: 'solidly',
+    zapAddress: '0x9da5A072804eE44B6259181BdE3F5393aEEf2bEC',
+    ammRouter: '0xa132DAB612dB5cB9fC9Ac426A0Cc215A3423F9c9',
+    ammFactory: '0x25CbdDb98b35ab1FF77413456B31EC81A6B6B746',
+    ammPairInitHash: '0xc1ac28b1c4ebe53c0cff67bab5878c4eb68759bb1e9f73977cd266b247d149f0',
+    lpProviderFee: 0.0002,
+  },
 ];
