+++ src/features/data/actions/deposit.ts
@@ -3,12 +3,13 @@ import { BeefyState } from '../../../redux-types';
 import { TokenAllowance } from '../apis/allowance/allowance-types';
 import { FetchAllBalancesResult } from '../apis/balance/balance-types';
 import { getAllowanceApi, getBalanceApi } from '../apis/instances';
-import { getEligibleZapOptions, ZapOptions } from '../apis/zap';
+import { getEligibleZapOptions } from '../apis/zap/zap';
 import { isTokenErc20, TokenEntity } from '../entities/token';
 import { isGovVault, isStandardVault, VaultEntity } from '../entities/vault';
 import { selectChainById } from '../selectors/chains';
 import { selectTokenByAddress } from '../selectors/tokens';
 import { selectVaultById } from '../selectors/vaults';
+import { ZapOptions } from '../apis/zap/zap-types';
 
 interface InitDepositFormParams {
   vaultId: VaultEntity['id'];
