+++ src/features/data/apis/config-types.ts
@@ -119,6 +119,7 @@ export interface ZapConfig {
   ammRouter: string;
   ammFactory: string;
   ammPairInitHash: string;
+  type?: 'uniswapv2' | 'solidly';
   withdrawEstimateMode?: 'getAmountOut' | 'getAmountsOut' | 'getAmountOutWithFee';
   withdrawEstimateFee?: string;
   lpProviderFee: number;
