+++ src/features/data/apis/zap/index.ts
@@ -1,0 +1,1 @@
+export * from './zap';
