+++ src/features/vault/components/FeeBreakdown/FeeBreakdown.tsx
@@ -5,7 +5,6 @@ import { styles } from './styles';
 import { Loader } from '../../../../components/Loader';
 import { BifiMaxis } from './BifiMaxis';
 import { isGovVault, VaultEntity } from '../../../data/entities/vault';
-import { ZapEstimate } from '../../../data/apis/zap';
 import { selectTokenByAddress } from '../../../data/selectors/tokens';
 import { useAppSelector } from '../../../../store';
 import { InterestTooltipContent } from '../../../home/components/Vault/components/InterestTooltipContent';
@@ -13,6 +12,7 @@ import { IconWithTooltip } from '../../../../components/Tooltip';
 import clsx from 'clsx';
 import { formatPercent } from '../../../../helpers/format';
 import { AlertWarning } from '../../../../components/Alerts';
+import { ZapEstimate } from '../../../data/apis/zap/zap-types';
 
 const useStyles = makeStyles(styles);
